Rewrite BentoProvider and useComponentsShowcase
